requirements.txt
torch
transformers
accelerate
nltk
pandas
sentencepiece
safetensors
pypdf
openai
google-genai
